set Time() EarthLocation(), and thus AltAz() frame
Here, the time and date of observation are saved as a string variable, and then the EarthLocation() of ARROW is set up (need longitude, latitude, and elevation). Once these are done, the AltAz observation frame is set up using AltAz(), with the obstime= and location= arguments set to the date time string and ARROW location.
Then, an if statement is used for *each* altaz coord to determine is the azimuth and altitude are not in the exclusion zone.